Move new tests to GroupTest.cs
